condensed unit-test
