--- test/routes/location-1.js
@@ -1,52 +1,0 @@
-'use strict'
-
-const Hapi = require('@hapi/hapi')
-const Lab = require('@hapi/lab')
-const Code = require('@hapi/code')
-const sinon = require('sinon')
-const lab = exports.lab = Lab.script()
-
-lab.experiment('Routes test - location - 1', () => {
-  let sandbox
-  let server
-
-  // Use a Sinon sandbox to manage spies, stubs and mocks for each test.
-  lab.beforeEach(async () => {
-    sandbox = await sinon.createSandbox()
-    server = Hapi.server({
-      port: 3000,
-      host: 'localhost'
-    })
-  })
-
-  lab.afterEach(async () => {
-    await server.stop()
-    await sandbox.restore()
-  })
-
-  lab.test('GET /location with no query parameters', async () => {
-    const locationPlugin = {
-      plugin: {
-        name: 'location',
-        register: (server, options) => {
-          server.route(require('../../server/routes/location'))
-        }
-      }
-    }
-
-    await server.register(locationPlugin)
-
-    await server.initialize()
-
-    // Tests JOI validation
-
-    const options = {
-      method: 'GET',
-      url: '/location'
-    }
-    const response = await server.inject(options)
-    const payload = JSON.parse(response.payload)
-    Code.expect(response.statusCode).to.equal(400)
-    Code.expect(payload.message).to.equal('Invalid request query input')
-  })
-})
